Treinamento 20-02-2024
Correção de exercícios e inicio do estudo da aula 08

+++ Curso_Python/Mundo_1/ex_15.py
@@ -1,7 +1,7 @@
 print("-=" * 35)
 print("Cáculo do Aluguel")
 print("-=" * 35)
-km = int(input("Quantos Km Rodados: "))
+km = float(input("Quantos Km Rodados: "))
 dias_usado = int(input("Quantos dias Alugados: "))
 valor_diario = 60
 valor_km = 0.15
